@@ -1,6 +1,6 @@
 {
   "name": "@sensenet/document-viewer-react",
-  "version": "1.0.0-development.1",
+  "version": "1.0.0",
   "description": "",
   "main": "dist/index.js",
   "files": [
@@ -18,12 +18,11 @@
     "commit": "git-cz",
     "pretest": "tslint --project tsconfig.test.json && rimraf temp && tsc -p tsconfig.test.json",
     "test": "rimraf coverage && nyc mocha -p tsconfig.test.json ./temp/test/index.js",
-    "check-coverage": "istanbul check-coverage ",
-    "report-coverage": "cat ./coverage/lcov.info | ./node_modules/.bin/codecov",
-    "semantic-release": "semantic-release pre && semantic-release post",
     "tslint": "tslint --project tsconfig.json",
     "prebuild": "npm run tslint",
-    "build": "rimraf dist && rimraf bundle && tsc -p tsconfig.json && webpack",
+    "build": "npm run build:tsc && npm run build:webpack",
+    "build:tsc": "rimraf dist && tsc -p tsconfig.json",
+    "build:webpack": "rimraf bundle && webpack",
     "publish:development": "npm run build && npm t && npm run typedoc && npm publish --tag development",
     "prepublishOnly": "npm run test && npm run build",
     "typedoc": "typedoc --mode file --out documentation src --tsconfig tsconfig.json --theme c:/Users/%USERNAME%/AppData/Roaming/npm/node_modules/@sensenet/typedoc-theme/sn-theme",
@@ -58,46 +57,44 @@
   },
   "homepage": "https://sensenet.com",
   "devDependencies": {
-    "@types/chai": "^4.1.3",
-    "@types/mocha": "^5.2.0",
-    "@types/react": "^16.3.13",
-    "@types/react-dom": "^16.0.5",
-    "@types/react-redux": "^5.0.18",
-    "@types/react-router-dom": "^4.2.6",
+    "@types/chai": "^4.1.4",
+    "@types/lodash": "^4.14.115",
+    "@types/mocha": "^5.2.5",
+    "@types/react": "^16.4.7",
+    "@types/react-dom": "^16.0.6",
+    "@types/react-fontawesome": "^1.6.3",
+    "@types/react-redux": "^6.0.5",
+    "@types/react-router-dom": "^4.3.0",
     "@types/react-scroll": "^1.5.3",
+    "@types/react-test-renderer": "^16.0.1",
     "@types/redux-logger": "^3.0.6",
-    "awesome-typescript-loader": "^5.0.0",
+    "@types/uuid": "^3.4.3",
+    "awesome-typescript-loader": "^5.2.0",
     "chai": "^4.1.2",
-    "codecov": "^3.0.1",
-    "copy-webpack-plugin": "^4.5.1",
-    "mocha": "^5.1.1",
-    "nyc": "^11.7.1",
+    "copy-webpack-plugin": "^4.5.2",
+    "mocha": "^5.2.0",
+    "nyc": "^12.0.2",
     "source-map-loader": "^0.2.3",
-    "tslint": "^5.9.1",
-    "typescript": "^2.8.3",
-    "webpack": "^4.6.0",
-    "webpack-bundle-analyzer": "^2.11.1",
-    "webpack-cli": "^2.0.15",
-    "webpack-dev-server": "^3.1.3"
+    "tslint": "^5.11.0",
+    "typescript": "^3.0.1",
+    "webpack": "^4.16.3",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-cli": "^2.1.5",
+    "webpack-dev-server": "^3.1.5"
   },
   "dependencies": {
-    "@types/lodash": "^4.14.106",
-    "@types/react-fontawesome": "^1.6.2",
-    "@types/react-test-renderer": "^16.0.1",
-    "@types/uuid": "^3.4.3",
+    "@material-ui/core": "^1.4.2",
+    "@material-ui/icons": "^1.1.0",
     "lodash.debounce": "^4.0.8",
-    "material-ui": "^1.0.0-beta.43",
-    "material-ui-icons": "^1.0.0-beta.36",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "react-fontawesome": "^1.6.1",
     "react-redux": "^5.0.7",
-    "react-scroll": "1.7.7",
-    "react-test-renderer": "^16.3.1",
-    "redux": "^3.7.2",
+    "react-test-renderer": "^16.4.1",
+    "redux": "^4.0.0",
+    "redux-di-middleware": "^1.0.2",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0",
-    "uuid": "^3.2.1"
+    "uuid": "^3.3.2"
   },
   "config": {
     "commitizen": {
