@@ -1,9 +1,10 @@
+import { SlideProps } from '@material-ui/core/Slide'
 import * as React from 'react'
 import { connect } from 'react-redux'
-import { ActionCreator, Dispatch } from 'redux'
-import { DocumentViewerSettings } from '../models'
+import { Action } from 'redux'
+import { InjectableAction } from 'redux-di-middleware'
 import { componentType } from '../services'
-import { DocumentStateType, pollDocumentData, RootReducerType } from '../store'
+import { pollDocumentData, RootReducerType } from '../store'
 import { LocalizationStateType, setLocalization } from '../store/Localization'
 import { DocumentViewerError } from './DocumentViewerError'
 import { DocumentViewerLayout } from './DocumentViewerLayout'
@@ -17,6 +18,8 @@ export interface OwnProps {
     documentIdOrPath: string | number
     version?: string
     localization?: Partial<LocalizationStateType>
+    drawerSlideProps?: Partial<SlideProps>
+
 }
 
 /**
@@ -38,7 +41,7 @@ const mapStateToProps = (state: RootReducerType, ownProps: OwnProps) => {
  * @param state the redux state
  */
 const mapDispatchToProps = {
-    pollDocumentData: pollDocumentData as ActionCreator<(dispatch: Dispatch<DocumentStateType>, getState: () => DocumentStateType, extraArgument: DocumentViewerSettings) => Promise<void>>,
+    pollDocumentData: pollDocumentData as (hostName: string, idOrPath: string | number, version?: string) => InjectableAction<RootReducerType, Action>,
     setLocalization,
 }
 
@@ -79,14 +82,14 @@ class DocumentViewer extends React.Component<docViewerComponentType> {
      * renders the component
      */
     public render() {
+        if (this.props.isLoading) {
+            return <DocumentViewerLoading />
+        }
+
         if (this.props.docViewerError || this.props.previewImagesError) {
             return <DocumentViewerError error={this.props.docViewerError || this.props.previewImagesError} />
         }
-        if (this.props.isLoading) {
-            return <DocumentViewerLoading />
-
-        }
-        return (<DocumentViewerLayout>
+        return (<DocumentViewerLayout drawerSlideProps={this.props.drawerSlideProps}>
             {this.props.children}
         </DocumentViewerLayout>)
     }
