@@ -1,9 +1,9 @@
-import { Drawer } from 'material-ui'
+import { Drawer } from '@material-ui/core'
+import { SlideProps } from '@material-ui/core/Slide'
 import React = require('react')
 import { connect } from 'react-redux'
-import { scroller } from 'react-scroll'
 import { componentType } from '../services'
-import { RootReducerType, setActivePages } from '../store'
+import { RootReducerType, setActivePages, setThumbnails } from '../store'
 import { PageList } from './'
 
 /**
@@ -15,6 +15,9 @@ const mapStateToProps = (state: RootReducerType) => {
         activePages: state.sensenetDocumentViewer.viewer.activePages,
         zoomMode: state.sensenetDocumentViewer.viewer.zoomMode,
         customZoomLevel: state.sensenetDocumentViewer.viewer.customZoomLevel,
+        showThumbnails: state.sensenetDocumentViewer.viewer.showThumbnails,
+        fitRelativeZoomLevel: state.sensenetDocumentViewer.viewer.fitRelativeZoomLevel,
+
     }
 }
 
@@ -24,82 +27,81 @@ const mapStateToProps = (state: RootReducerType) => {
  */
 const mapDispatchToProps = {
     setActivePages,
+    setThumbnails,
+}
+
+/** Props definition for the Document Viewer layout */
+export interface DocumentLayoutOwnProps {
+    drawerSlideProps?: Partial<SlideProps>
 }
 
 /** State type definition for the DocumentViewerLayout component */
 export interface DocumentLayoutState {
-    showThumbnails: boolean
     activePage?: number
+    thumbnaislVisibility: boolean
 }
 
 /**
  * Component for the main DocumentViewer layout
  */
-class DocumentViewerLayoutComponent extends React.Component<componentType<typeof mapStateToProps, typeof mapDispatchToProps, undefined>, DocumentLayoutState> {
+class DocumentViewerLayoutComponent extends React.Component<componentType<typeof mapStateToProps, typeof mapDispatchToProps, DocumentLayoutOwnProps>, DocumentLayoutState> {
 
     /** the component state */
-    public state = { showThumbnails: true, activePage: 1 }
+    public state = { activePage: 1, thumbnaislVisibility: this.props.showThumbnails }
 
     /** scrolls the viewer to focus to the page with the provided index */
-    public scrollTo(index: number) {
+    public scrollTo(index: number, smoothScroll: boolean = true) {
         this.setState({ ...this.state, activePage: index }, () => {
-            scroller.scrollTo(`Page-${index}`, {
-                containerId: 'sn-document-viewer-pages',
-                smooth: 'easeInOutQuint',
-                duration: 600,
-                offset: -8,
-            })
+            const pagesContainer = document.getElementById('sn-document-viewer-pages')
+            const activePage = document.getElementById(`Page-${index}`)
 
-            if (this.state.showThumbnails) {
-                scroller.scrollTo(`Thumbnail-${index}`, {
-                    containerId: 'sn-document-viewer-thumbnails',
-                    smooth: 'easeInOutQuint',
-                    duration: 600,
-                    offset: -8,
+            if (pagesContainer && pagesContainer.scrollTo && activePage) {
+                pagesContainer.scrollTo({
+                    top: activePage.offsetTop - 8,
+                    behavior: smoothScroll ? 'smooth' : 'instant',
+                })
+            }
+
+            const thumbnailsContainer = document.getElementById('sn-document-viewer-thumbnails')
+            const activeThumbnail = document.getElementById(`Thumbnail-${index}`)
+
+            if (thumbnailsContainer && thumbnailsContainer.scrollTo && activeThumbnail) {
+                thumbnailsContainer.scrollTo({
+                    top: activeThumbnail.offsetTop - 16,
+                    behavior: 'smooth',
                 })
             }
 
             if (this.props.activePages[0] !== index) {
                 this.props.setActivePages([index])
             }
-
         })
-
     }
 
     /** triggered when the component will receive props */
     public componentWillReceiveProps(newProps: this['props']) {
-        if (this.props.activePages[0] !== newProps.activePages[0]) {
-            this.scrollTo(newProps.activePages[0])
+        if (this.props.activePages[0] !== newProps.activePages[0] || this.props.fitRelativeZoomLevel !== newProps.fitRelativeZoomLevel) {
+            this.scrollTo(newProps.activePages[0], this.props.fitRelativeZoomLevel === newProps.fitRelativeZoomLevel)
         }
-    }
+        if (this.props.showThumbnails !== newProps.showThumbnails) {
+            if (newProps.showThumbnails) {
+                this.setState({
+                    ...this.state,
+                    thumbnaislVisibility: true,
+                })
+                window.dispatchEvent(new Event('resize'))
+            } else {
+                setTimeout(() => {
+                    this.setState({
+                        ...this.state,
+                        thumbnaislVisibility: false,
+                    })
+                    window.dispatchEvent(new Event('resize'))
+                }, 200)
+            }
 
-    private onResize() {
-        const showThumbnails = innerWidth > 800
-        if (this.state.showThumbnails !== showThumbnails) {
-            this.setState({
-                ...this.state,
-                showThumbnails,
-            })
         }
-    }
 
-    private resizeWatcher = this.onResize.bind(this)
-
-    /** Event before the component will mount */
-    public componentWillMount() {
-        addEventListener('resize', this.resizeWatcher)
-    }
-
-    /** Event after the component did mount */
-    public componentDidMount() {
-        this.onResize()
-
-    }
-
-    /** Event before the component will unmount */
-    public componentWillUnmount() {
-        removeEventListener('resize', this.resizeWatcher)
     }
 
     /**
@@ -122,11 +124,33 @@ class DocumentViewerLayoutComponent extends React.Component<componentType<typeof
                     zIndex: 0,
                     position: 'relative',
                 }}>
+                    <Drawer
+                        variant={'persistent'}
+                        open={this.props.showThumbnails}
+                        anchor="left"
+                        SlideProps={this.props.drawerSlideProps}
+                        PaperProps={{ style: { position: 'relative', width: this.state.thumbnaislVisibility ? '200px' : 0, height: '100%', overflow: 'hidden' } }}>
+                        <PageList
+                            showWidgets={false}
+                            style={{ minWidth: 200, marginRight: '-16px', paddingRight: 0 }}
+                            id="sn-document-viewer-thumbnails"
+                            zoomMode="fit"
+                            fitRelativeZoomLevel={0}
+                            zoomLevel={1}
+                            onPageClick={(ev, index) => this.scrollTo(index)}
+                            elementNamePrefix="Thumbnail-"
+                            images="thumbnail"
+                            tolerance={0}
+                            padding={16}
+                            activePage={this.state.activePage}
+                        />
+                    </Drawer>
                     <PageList
                         showWidgets={true}
                         id="sn-document-viewer-pages"
                         zoomMode={this.props.zoomMode}
                         zoomLevel={this.props.customZoomLevel}
+                        fitRelativeZoomLevel={this.props.fitRelativeZoomLevel}
                         onPageClick={(ev, index) => this.scrollTo(index)}
                         elementNamePrefix="Page-"
                         images="preview"
@@ -134,25 +158,6 @@ class DocumentViewerLayoutComponent extends React.Component<componentType<typeof
                         padding={8}
                         activePage={this.state.activePage}
                     />
-
-                    {this.state.showThumbnails ?
-                        <Drawer variant={'persistent'} open anchor="right" PaperProps={{ style: { position: 'relative', width: '200px', height: '100%' } }}>
-                            <PageList
-                                showWidgets={false}
-                                style={{ minWidth: 160 }}
-                                id="sn-document-viewer-thumbnails"
-                                zoomMode="fit"
-                                zoomLevel={1}
-                                onPageClick={(ev, index) => this.scrollTo(index)}
-                                elementNamePrefix="Thumbnail-"
-                                images="thumbnail"
-                                tolerance={0}
-                                padding={8}
-                                activePage={this.state.activePage}
-                            />
-                        </Drawer>
-                        : null
-                    }
                 </div>
             </div >
         )
