@@ -1,15 +1,14 @@
-import { Paper } from 'material-ui'
-import { CircularProgress } from 'material-ui/Progress'
+import { Paper } from '@material-ui/core'
+import { CircularProgress } from '@material-ui/core'
 import React = require('react')
 import { connect } from 'react-redux'
-import { Element } from 'react-scroll'
-import { DocumentData, DocumentViewerSettings, PreviewImageData } from '../models'
+import { DocumentData, PreviewImageData } from '../models'
 import { componentType, ImageUtil } from '../services'
 import { previewAvailable, RootReducerType, ZoomMode } from '../store'
 
-import { ActionCreator } from 'redux'
-import { ThunkAction } from 'redux-thunk'
-import { RotatePageWidget, ShapesWidget } from './page-widgets'
+import { Action } from 'redux'
+import { InjectableAction } from 'redux-di-middleware'
+import { ShapesWidget } from './page-widgets'
 
 /**
  * maps state fields from the store to component props
@@ -17,7 +16,6 @@ import { RotatePageWidget, ShapesWidget } from './page-widgets'
  */
 const mapStateToProps = (state: RootReducerType, ownProps: { imageIndex: number }) => {
     return {
-        store: state,
         documentData: state.sensenetDocumentViewer.documentState.document as DocumentData,
         version: state.sensenetDocumentViewer.documentState.version,
         page: state.sensenetDocumentViewer.previewImages.AvailableImages[ownProps.imageIndex - 1] || {} as PreviewImageData,
@@ -32,7 +30,7 @@ const mapStateToProps = (state: RootReducerType, ownProps: { imageIndex: number 
  * @param state the redux state
  */
 const mapDispatchToProps = {
-    previewAvailable: previewAvailable as ActionCreator<ThunkAction<Promise<void>, RootReducerType, DocumentViewerSettings>>,
+    previewAvailable: previewAvailable as (documentData: DocumentData, version?: string, page?: number) => InjectableAction<RootReducerType, Action>,
 }
 
 /**
@@ -46,7 +44,9 @@ export interface OwnProps {
     elementNamePrefix: string,
     zoomMode: ZoomMode,
     zoomLevel: number,
+    fitRelativeZoomLevel: number,
     onClick: (ev: React.MouseEvent<HTMLElement>) => any,
+    margin: number,
     image: 'preview' | 'thumbnail'
 }
 
@@ -62,13 +62,19 @@ export interface PageState {
     imageHeight: string
     imageTransform: string
     zoomRatio: number
+    isPolling: boolean
 }
 
 class Page extends React.Component<componentType<typeof mapStateToProps, typeof mapDispatchToProps, OwnProps>, PageState> {
 
-    private pollPreview?: number
+    private pollPreview?: number = setInterval(() => {
+        if (this.state.isPolling) {
+            this.props.previewAvailable(this.props.documentData, this.props.version, this.props.imageIndex)
+        }
+    }, this.props.pollInterval) as any as number
+
     /** the component state */
-    public state = this.getStateFromProps(this.props)
+    public state = Page.getDerivedStateFromProps(this.props)
 
     private stopPolling() {
         if (this.pollPreview) {
@@ -77,20 +83,19 @@ class Page extends React.Component<componentType<typeof mapStateToProps, typeof 
         }
     }
 
-    /** event after the component did mount */
-    public componentDidMount() {
-        this.componentWillReceiveProps(this.props)
-    }
-
     /** event before the component did unmount */
     public componentWillUnmount() {
         this.stopPolling()
     }
 
-    private getStateFromProps(props: this['props']): PageState {
+    /**
+     * Returns a derived state from the specified props
+     * @param props The props for state creation
+     */
+    public static getDerivedStateFromProps(props: Page['props']): PageState {
         const imageRotation = ImageUtil.normalizeDegrees(props.page.Attributes && props.page.Attributes.degree || 0)
         const imageRotationRads = (imageRotation % 180) * Math.PI / 180
-        const imgSrc = (this.props.image === 'preview' ? props.page.PreviewImageUrl : props.page.ThumbnailImageUrl) || ''
+        const imgSrc = (props.image === 'preview' ? props.page.PreviewImageUrl : props.page.ThumbnailImageUrl) || ''
         const relativePageSize = ImageUtil.getImageSize({
             width: props.viewportWidth,
             height: props.viewportHeight,
@@ -98,7 +103,7 @@ class Page extends React.Component<componentType<typeof mapStateToProps, typeof 
                 width: props.page.Width,
                 height: props.page.Height,
                 rotation: props.page.Attributes && props.page.Attributes.degree || 0,
-            }, props.zoomMode, props.zoomLevel)
+            }, props.zoomMode, props.zoomLevel, props.fitRelativeZoomLevel)
         const boundingBox = ImageUtil.getRotatedBoundingBoxSize({
             width: props.page.Width,
             height: props.page.Height,
@@ -107,17 +112,8 @@ class Page extends React.Component<componentType<typeof mapStateToProps, typeof 
         const maxDiff = (relativePageSize.height - relativePageSize.width) / 2
         const diffHeight = Math.sin(imageRotationRads) * maxDiff
 
-        if (!imgSrc) {
-            this.stopPolling()
-            this.pollPreview = setInterval(() => {
-                this.props.previewAvailable(this.props.documentData, this.props.version, this.props.imageIndex)
-            }, this.props.pollInterval) as any as number
-        } else {
-            this.stopPolling()
-        }
-
         return {
-            isActive: props.activePages.indexOf(this.props.page.Index) >= 0,
+            isActive: props.activePages.indexOf(props.page.Index) >= 0,
             imgSrc,
             pageWidth: relativePageSize.width,
             pageHeight: relativePageSize.height,
@@ -125,47 +121,39 @@ class Page extends React.Component<componentType<typeof mapStateToProps, typeof 
             imageWidth: `${100 * boundingBox.zoomRatio}%`,
             imageHeight: `${100 * boundingBox.zoomRatio}%`,
             imageTransform: `translateY(${diffHeight}px) rotate(${imageRotation}deg)`,
+            isPolling: !(imgSrc ? true : false),
         }
-    }
-
-    /** triggered when the component will receive props */
-
-    public componentWillReceiveProps(nextProps: this['props']) {
-        const newState = this.getStateFromProps(nextProps)
-        this.setState({ ...newState })
     }
 
     /**
      * renders the component
      */
     public render() {
+        const elementName = `${this.props.elementNamePrefix}${this.props.page.Index}`
         return (
-            <Element name={`${this.props.elementNamePrefix}${this.props.page.Index}`} style={{ margin: '8px' }}>
-                <Paper elevation={this.state.isActive ? 8 : 2}>
-                    <div style={{
-                        padding: 0,
-                        overflow: 'hidden',
-                        width: this.state.pageWidth,
-                        height: this.state.pageHeight,
-                        position: 'relative',
-                    }} onClick={(ev) => this.props.onClick(ev)}>
-                        {this.props.showWidgets ?
-                            <div>
-                                <ShapesWidget zoomRatio={this.state.zoomRatio} page={this.props.page} viewPort={{ height: this.state.pageHeight, width: this.state.pageWidth }} />
-                                <RotatePageWidget zoomRatio={this.state.zoomRatio} page={this.props.page} viewPort={{ height: this.state.pageHeight, width: this.state.pageWidth }} />
-                            </div>
-                            : null}
-                        <span style={{ display: 'flex', justifyContent: 'center' }}>
-                            {this.state.imgSrc ?
-                                <img src={`${this.state.imgSrc}${this.props.showWatermark ? '?watermark=true' : ''}`}
-                                    style={{ transition: 'transform .1s ease-in-out', width: this.state.imageWidth, height: this.state.imageHeight, transform: this.state.imageTransform }}
-                                /> :
-                                <CircularProgress />
-                            }
-                        </span>
-                    </div>
-                </Paper>
-            </Element >
+            <Paper elevation={this.state.isActive ? 8 : 2} id={elementName} style={{ margin: this.props.margin }}>
+                <div style={{
+                    padding: 0,
+                    overflow: 'hidden',
+                    width: this.state.pageWidth - 2 * this.props.margin,
+                    height: this.state.pageHeight - 2 * this.props.margin,
+                    position: 'relative',
+                }} onClick={(ev) => this.props.onClick(ev)}>
+                    {this.props.showWidgets ?
+                        <div>
+                            <ShapesWidget zoomRatio={this.state.zoomRatio} page={this.props.page} viewPort={{ height: this.state.pageHeight, width: this.state.pageWidth }} />
+                        </div>
+                        : null}
+                    <span style={{ display: 'flex', justifyContent: 'center' }}>
+                        {this.state.imgSrc ?
+                            <img src={`${this.state.imgSrc}${this.props.showWatermark ? '?watermark=true' : ''}`}
+                                style={{ transition: 'transform .1s ease-in-out', width: this.state.imageWidth, height: this.state.imageHeight, transform: this.state.imageTransform }}
+                            /> :
+                            <CircularProgress style={{ marginTop: '50%' }} />
+                        }
+                    </span>
+                </div>
+            </Paper>
         )
     }
 }
