@@ -1,5 +1,5 @@
-import { Divider, IconButton, Menu, MenuItem, MobileStepper } from 'material-ui'
-import { AspectRatio, Code, Error, ZoomIn, ZoomOut, ZoomOutMap } from 'material-ui-icons'
+import { Divider, IconButton, Menu, MenuItem, MobileStepper } from '@material-ui/core'
+import { AspectRatio, Code, Error, ZoomIn, ZoomOut, ZoomOutMap } from '@material-ui/icons'
 import * as React from 'react'
 import { connect } from 'react-redux'
 import { componentType } from '../../services'
