@@ -1,7 +1,14 @@
-export { PagerWidget } from './Pager'
+export { DocumentTitlePager } from './DocumentTitlePager'
+export { Download } from './DownloadWidget'
+export { Print } from './PrintWidget'
+export { Share } from './ShareWidget'
+export { RotateActivePages } from './RotateActivePages'
 export { RotateDocumentWidget } from './RotateDocument'
 export { SaveWidget } from './SaveWidget'
+export { SearchBar } from './SearchBar'
 export { ToggleRedactionWidget } from './ToggleRedaction'
 export { ToggleShapesWidget } from './ToggleShapes'
+export { ToggleThumbnailsWidget } from './ToggleThumbnails'
 export { ToggleWatermarkWidget } from './ToggleWatermark'
+export { ZoomInOutWidget } from './ZoomInOut'
 export { ZoomModeWidget } from './ZoomMode'
