@@ -12,7 +12,7 @@ const storeConfig = getStoreConfig(exampleSettings)
 
 const rootReducer = combineReducers({
     sensenetDocumentViewer: sensenetDocumentViewerReducer,
-    myStuff: (state = {}, action) => {
+    myStuff: (state = {}, action: any) => {
         if (action.type === 'SN_DOCVEWER_DOCUMENT_SAVE_CHANGES_SUCCESS') {
             // tslint:disable-next-line:no-console
             // console.log('Changes saved')
