@@ -1,5 +1,5 @@
-import { ActionCreator, Reducer } from 'redux'
-import { ThunkAction } from 'redux-thunk'
+import { Action, Reducer } from 'redux'
+import { InjectableAction } from 'redux-di-middleware'
 import { PreviewState } from '../Enums'
 import { DocumentData, DocumentViewerSettings, PreviewImageData, Shape, Shapes } from '../models'
 import { Dimensions, ImageUtil } from '../services'
@@ -23,39 +23,50 @@ export interface DocumentStateType {
 }
 
 /**
+ * Resets the document viewer state to the default.
+ */
+export const resetDocumentData = () => ({
+    type: 'SN_DOCVIEWER_DOCUMENT_RESET_DOCUMENT',
+})
+
+/**
  * Thunk action that polls document data from the specified API
  * @param hostName the host name, e.g. 'https://my-sensenet-site.com'
  * @param idOrPath Id or full path for the document, e.g.: 'Root/Sites/MySite/MyDocLib/('doc.docx')
  * @param version The document version
  */
-export const pollDocumentData: ActionCreator<ThunkAction<Promise<void>, DocumentStateType, DocumentViewerSettings>> = (hostName: string, idOrPath: string, version: string) => {
-    return async (dispatch, getState, api) => {
-        let docData: DocumentData | undefined
-        while (!docData || docData.pageCount === PreviewState.Loading) {
-            try {
-                docData = await api.getDocumentData({ idOrPath, hostName })
-                if (!docData || docData.pageCount === PreviewState.Loading) {
-                    await new Promise<void>((resolve, reject) => setTimeout(() => { resolve() }, getState().pollInterval))
+export const pollDocumentData: (hostName: string, idOrPath: string | number, version?: string) => InjectableAction<RootReducerType, Action> =
+    (hostName: string, idOrPath: string | number, version: string = 'V1.0A') => ({
+        type: 'SN_POLL_DOCUMENT_DATA_INJECTABLE_ACTION',
+        inject: async (options) => {
+            const api = options.getInjectable(DocumentViewerSettings)
+            options.dispatch(resetDocumentData())
+            let docData: DocumentData | undefined
+            while (!docData || docData.pageCount === PreviewState.Loading) {
+                try {
+                    docData = await api.getDocumentData({ idOrPath, hostName })
+                    if (!docData || docData.pageCount === PreviewState.Loading) {
+                        await new Promise<void>((resolve) => setTimeout(() => { resolve() }, options.getState().sensenetDocumentViewer.documentState.pollInterval))
+                    }
+                } catch (error) {
+                    options.dispatch(documentReceiveErrorAction(error || Error('Error loading document')))
+                    return
                 }
-            } catch (error) {
-                dispatch(documentReceiveErrorAction(error || Error('Error loading document')))
-                return
             }
-        }
-        try {
-            const [canEdit, canHideRedaction, canHideWatermark] = await Promise.all([
-                await api.canEditDocument(docData),
-                await api.canHideRedaction(docData),
-                await api.canHideWatermark(docData),
-            ])
-            dispatch(documentPermissionsReceived(canEdit, canHideRedaction, canHideWatermark))
-        } catch (error) {
-            dispatch(documentPermissionsReceived(false, false, false))
-        }
-        dispatch(documentReceivedAction(docData))
-        dispatch<any>(getAvailableImages(docData))
-    }
-}
+            try {
+                const [canEdit, canHideRedaction, canHideWatermark] = await Promise.all([
+                    await api.canEditDocument(docData),
+                    await api.canHideRedaction(docData),
+                    await api.canHideWatermark(docData),
+                ])
+                options.dispatch(documentPermissionsReceived(canEdit, canHideRedaction, canHideWatermark))
+            } catch (error) {
+                options.dispatch(documentPermissionsReceived(false, false, false))
+            }
+            options.dispatch(documentReceivedAction(docData))
+            options.dispatch<any>(getAvailableImages(docData))
+        },
+    })
 
 /**
  * Action that updates the store with the received document data
@@ -70,7 +81,7 @@ export const documentReceivedAction = (document: DocumentData) => ({
  * Action that updates the store with a document receive error
  * @param error The error message
  */
-export const documentReceiveErrorAction = (error: string) => ({
+export const documentReceiveErrorAction = (error: any) => ({
     type: 'SN_DOCVIEWER_DOCUMENT_DATA_RECEIVE_ERROR',
     error,
 })
@@ -115,11 +126,11 @@ export const setPollInterval = (pollInterval: number) => ({
  * @param canHideWatermark 'Can hide watermark' permission value
  */
 export const documentPermissionsReceived = (canEdit: boolean, canHideRedaction: boolean, canHideWatermark: boolean) => ({
-        type: 'SN_DOCVEWER_DOCUMENT_PERMISSIONS_RECEIVED',
-        canEdit,
-        canHideRedaction,
-        canHideWatermark,
-    })
+    type: 'SN_DOCVEWER_DOCUMENT_PERMISSIONS_RECEIVED',
+    canEdit,
+    canHideRedaction,
+    canHideWatermark,
+})
 
 /**
  * Action that will be fired when a save request has been sent
@@ -158,17 +169,19 @@ export const rotateShapesForPages = (pages: Array<{ index: number, size: Dimensi
 /**
  * Thunk action to call the Save endpoint with the current document state to save changes
  */
-export const saveChanges: ActionCreator<ThunkAction<Promise<void>, RootReducerType, DocumentViewerSettings>> = () => {
-    return async (dispatch, getState, api) => {
-        dispatch(saveChangesRequest())
+export const saveChanges: () => InjectableAction<RootReducerType, Action> = () => ({
+    type: 'SN_DOCVIEWER_SAVE_CHANGES_INJECTABLE_ACTION',
+    inject: async (options) => {
+        const api = options.getInjectable(DocumentViewerSettings)
+        options.dispatch(saveChangesRequest())
         try {
-            await api.saveChanges(getState().sensenetDocumentViewer.documentState.document as DocumentData, getState().sensenetDocumentViewer.previewImages.AvailableImages as PreviewImageData[])
-            dispatch(saveChangesSuccess())
+            await api.saveChanges(options.getState().sensenetDocumentViewer.documentState.document as DocumentData, options.getState().sensenetDocumentViewer.previewImages.AvailableImages as PreviewImageData[])
+            options.dispatch(saveChangesSuccess())
         } catch (error) {
-            dispatch(saveChangesError(error))
+            options.dispatch(saveChangesError(error))
         }
-    }
-}
+    },
+})
 
 /**
  * helper method to apply shape rotations
@@ -197,43 +210,48 @@ export const applyShapeRotations = <T extends Shape>(shapes: T[], degree: number
 ])
 
 /**
+ * The default state data for the Document
+ */
+export const defaultState: DocumentStateType = {
+    document: {
+        hostName: '',
+        shapes: {
+            annotations: [],
+            highlights: [],
+            redactions: [],
+        },
+        documentName: '',
+        documentType: '',
+        fileSizekB: 0,
+        idOrPath: 0,
+        pageAttributes: [],
+        pageCount: -1,
+    } as DocumentData,
+    error: undefined,
+    isLoading: true,
+    idOrPath: undefined,
+    version: undefined,
+    canEdit: false,
+    hasChanges: false,
+    canHideRedaction: false,
+    canHideWatermark: false,
+    pollInterval: 2000,
+}
+
+/**
  * Reducer method for the document state
  * @param state the current state
  * @param action the action to dispatch
  */
 export const documentStateReducer: Reducer<DocumentStateType>
-    = (state = {
-        document: {
-            hostName: '',
-            shapes: {
-                annotations: [],
-                highlights: [],
-                redactions: [],
-            },
-            documentName: '',
-            documentType: '',
-            fileSizekB: 0,
-            idOrPath: 0,
-            pageAttributes: [],
-            pageCount: 0,
-        } as DocumentData,
-        error: undefined,
-        isLoading: true,
-        idOrPath: undefined,
-        version: undefined,
-        canEdit: false,
-        hasChanges: false,
-        canHideRedaction: false,
-        canHideWatermark: false,
-        pollInterval: 2000,
-    }, action) => {
+    = (state = defaultState, action) => {
         switch (action.type) {
-            case 'SN_DOCVIEWER_DOCUMENT_SET_DOCUMENT':
-                return { ...state, error: undefined, isLoading: true, idOrPath: action.idOrPath, version: action.version }
+            case 'SN_DOCVIEWER_DOCUMENT_RESET_DOCUMENT':
+                return { ...defaultState }
             case 'SN_DOCVIEWER_DOCUMENT_DATA_RECEIVED':
-                return { ...state, document: {...state.document, ...action.document}, error: undefined, isLoading: false, hasChanges: false }
+                return { ...state, document: { ...state.document, ...action.document }, error: undefined, isLoading: false, hasChanges: false }
             case 'SN_DOCVIEWER_DOCUMENT_DATA_RECEIVE_ERROR':
-                return { ...state, error: action.error, isLoading: false }
+                return { ...state, document: { ...state.document, pageCount: 0 }, error: action.error, isLoading: false }
             case 'SN_DOCVIEWER_DOCUMENT_UPDATE_SHAPE':
                 return {
                     ...state,
