@@ -1,4 +1,5 @@
 import { Reducer } from 'redux'
+import { PreviewState } from '../Enums'
 
 /**
  * Model definition for the localization store
@@ -17,6 +18,7 @@ export interface LocalizationStateType {
     toggleRedaction: string
     toggleWatermark: string
     toggleShapes: string
+    toggleThumbnails: string
     firstPage: string
     previousPage: string
     gotoPage: string
@@ -24,9 +26,13 @@ export interface LocalizationStateType {
     lastPage: string
     saveChanges: string
     loadingDocument: string
-    errorLoadingDocument: string
+    errorLoadingDocument: Array<{ state: PreviewState, message: string, details: string, code?: number }>
     errorLoadingDetails: string
     reloadPage: string
+    search: string
+    share: string
+    download: string
+    print: string
 }
 
 /**
@@ -46,16 +52,29 @@ export const defaultLocalization: LocalizationStateType = {
     toggleRedaction: 'Toggle redaction',
     toggleWatermark: 'Toggle watermark',
     toggleShapes: 'Toggle shapes',
+    toggleThumbnails: 'Toggle thumbnails',
     firstPage: 'First page',
     previousPage: 'Previous page',
     gotoPage: 'Goto page',
     nextPage: 'Next page',
     lastPage: 'Last page',
     saveChanges: 'Save changes',
-    loadingDocument: 'Loading document',
-    errorLoadingDocument: 'Error loading document',
+    loadingDocument: 'Preview image generation is in progress',
+    errorLoadingDocument: [
+        { code: 500, state: PreviewState.Empty, message: 'Ooops! Something went wrong...', details: 'An unexpected error seems to have occured. Why not try refreshing your page? If the page still not working try to check back later.' },
+        { code: 404, state: PreviewState.Empty, message: 'Ooops! Something went wrong...', details: `You don't have permission to see this document \r\n The document you are trying to open has been deleted \r\n The document has been moved to another place ` },
+        { state: PreviewState.UploadFailure, message: 'Failed to upload', details: '' },
+        { state: PreviewState.UploadFailure2, message: 'Failed to upload', details: '' },
+        { state: PreviewState.ExtensionFailure, message: 'Failed to generate preview images due to an extension error', details: '' },
+        { state: PreviewState.Empty, message: 'The document doesn\'t have any preview images', details: '' },
+        { state: PreviewState.NoPreviewProviderEnabled, message: 'There is no preview provider enabled', details: '' },
+    ],
     errorLoadingDetails: 'The following error occured during opening a document: ',
     reloadPage: 'Reload page',
+    search: 'Search',
+    share: 'Share',
+    download: 'Download',
+    print: 'Print',
 }
 
 /**
@@ -72,7 +91,7 @@ export const setLocalization = (localization: Partial<LocalizationStateType>) =>
  * @param state the current state
  * @param action the action to dispatch
  */
-export const localizationReducer: Reducer<LocalizationStateType> = (state: LocalizationStateType = defaultLocalization, action ) => {
+export const localizationReducer: Reducer<LocalizationStateType> = (state: LocalizationStateType = defaultLocalization, action) => {
     switch (action.type) {
         case 'SN_DOCVIEWER_SET_LOCALIZATION': {
             return {
