@@ -1,8 +1,8 @@
-import { Store } from 'react-redux'
-import { applyMiddleware, createStore } from 'redux'
+import { Store } from 'redux'
+import { Action, applyMiddleware, createStore } from 'redux'
 import { Reducer } from 'redux'
 import { StoreEnhancer } from 'redux'
-import thunk from 'redux-thunk'
+import { ReduxDiMiddleware } from 'redux-di-middleware'
 import { DocumentViewerSettings } from '../models'
 import { rootReducer, RootReducerType } from './RootReducer'
 
@@ -10,14 +10,16 @@ import { rootReducer, RootReducerType } from './RootReducer'
  * gets a configuration object for the Store instance
  * @param {DocumentViewerSettings} settings The Settings object for the document viewer instance
  */
-export const getStoreConfig: (settings: DocumentViewerSettings) => {rootReducer: Reducer<RootReducerType>, preloadedState: RootReducerType, enhancer: StoreEnhancer<any>}
-     = (settings: DocumentViewerSettings) => {
-    return {
-        rootReducer,
-        preloadedState: {sensenetDocumentViewer: {documentState: {isLoading: true}}} as RootReducerType,
-        enhancer: applyMiddleware(thunk.withExtraArgument(settings)),
+export const getStoreConfig: (settings: DocumentViewerSettings) => { rootReducer: Reducer<RootReducerType>, preloadedState: RootReducerType, enhancer: StoreEnhancer<any> }
+    = (settings: DocumentViewerSettings) => {
+        const di = new ReduxDiMiddleware()
+        di.setInjectable(settings)
+        return {
+            rootReducer,
+            preloadedState: { sensenetDocumentViewer: { documentState: { isLoading: true } } } as RootReducerType,
+            enhancer: applyMiddleware(di.getMiddleware()),
+        }
     }
-}
 
 /**
  * returns a Store object for a Document Viewer instance
@@ -25,7 +27,7 @@ export const getStoreConfig: (settings: DocumentViewerSettings) => {rootReducer:
  */
 export const configureStore: (settings: DocumentViewerSettings) => Store<RootReducerType> = (settings: DocumentViewerSettings) => {
     const config = getStoreConfig(settings)
-    return createStore<RootReducerType>(config.rootReducer, config.preloadedState, config.enhancer)
+    return createStore<RootReducerType, Action, {}, {}>(config.rootReducer, config.preloadedState, config.enhancer)
 }
 
 export * from './Document'
