@@ -9,12 +9,12 @@ import { exampleDocumentData, useTestContext, useTestContextWithSettings } from 
 /**
  * Tests for the Document Viewer main component
  */
-export const documentViewerTests = describe('Document Viewer component', () => {
+export const documentViewerTests: Mocha.Suite = describe('Document Viewer component', () => {
     it('Should render without crashing', () => {
         useTestContext((ctx) => {
             const c = renderer.create(
                 <Provider store={ctx.store} >
-                    <DocumentViewer documentIdOrPath="" hostName=""/>
+                    <DocumentViewer documentIdOrPath="" hostName="" drawerSlideProps={{ in: true }} />
                 </Provider>)
             c.unmount()
         })
@@ -27,7 +27,7 @@ export const documentViewerTests = describe('Document Viewer component', () => {
         })
         const exampleIdOrPath = 'Example/Id/Or/Path'
         useTestContextWithSettings({
-            getDocumentData: async ({idOrPath}) => {
+            getDocumentData: async ({ idOrPath }) => {
                 expect(idOrPath).to.be.eq(exampleIdOrPath)
                 done()
                 return exampleDocumentData
@@ -35,7 +35,7 @@ export const documentViewerTests = describe('Document Viewer component', () => {
         }, (ctx) => {
             c = renderer.create(
                 <Provider store={ctx.store}>
-                    <DocumentViewer documentIdOrPath={exampleIdOrPath} hostName="" />
+                    <DocumentViewer documentIdOrPath={exampleIdOrPath} hostName="" drawerSlideProps={{ in: true }} />
                 </Provider>)
         })
     })
@@ -43,27 +43,33 @@ export const documentViewerTests = describe('Document Viewer component', () => {
     it('Should start polling again on path change', (done: MochaDone) => {
         let c!: renderer.ReactTestRenderer
         let pollCount = 0
+        let promise!: Promise<void>
         after(() => {
             c.unmount()
         })
-        const exampleIdOrPath = 'Example/Id/Or/Path'
-        useTestContextWithSettings({
-            getDocumentData: async (idOrPath) => {
-                pollCount++
-                if (pollCount === 2) {
-                    done()
-                }
-                return exampleDocumentData
-            },
-        }, (ctx) => {
-            c = renderer.create(
-                <Provider store={ctx.store} >
-                    <DocumentViewer documentIdOrPath={exampleIdOrPath} hostName="" />
-                </Provider>)
-            setTimeout(() => {
-                const component = c.root.findByType(DocumentViewer).children[0] as renderer.ReactTestInstance
-                component.instance.componentWillReceiveProps({ ...component.instance.props, documentIdOrPath: 123456 })
-            }, 100)
+
+        promise = new Promise((resolve) => {
+            const exampleIdOrPath = 'Example/Id/Or/Path'
+            useTestContextWithSettings({
+                getDocumentData: async () => {
+                    pollCount++
+                    if (pollCount === 2) {
+                        resolve()
+                        done()
+                    }
+                    return exampleDocumentData
+                },
+            }, async (ctx) => {
+                c = renderer.create(
+                    <Provider store={ctx.store} >
+                        <DocumentViewer documentIdOrPath={exampleIdOrPath} hostName="" drawerSlideProps={{ in: true }} />
+                    </Provider>)
+                setTimeout(() => {
+                    const component = c.root.findByType(DocumentViewer).children[0] as renderer.ReactTestInstance
+                    component.instance.componentWillReceiveProps({ ...component.instance.props, documentIdOrPath: 123456 })
+                }, 100)
+                await promise
+            })
         })
     })
 
@@ -74,13 +80,11 @@ export const documentViewerTests = describe('Document Viewer component', () => {
         })
         const exampleIdOrPath = 'Example/Id/Or/Path'
         useTestContextWithSettings({
-            getDocumentData: async (idOrPath) => {
-                throw Error('Nooo')
-            },
+            getDocumentData: () => Promise.reject('Nooo'),
         }, (ctx) => {
             c = renderer.create(
                 <Provider store={ctx.store} >
-                    <DocumentViewer documentIdOrPath={exampleIdOrPath} hostName="" />
+                    <DocumentViewer documentIdOrPath={exampleIdOrPath} hostName="" drawerSlideProps={{ in: true }} />
                 </Provider>)
             setTimeout(() => {
                 const error = c.root.findByType(DocumentViewerError).instance
