@@ -9,7 +9,7 @@ import { exampleDocumentData, useTestContext } from '../viewercontext'
 /**
  * LayoutAppBar Component tests
  */
-export const layoutAppBarTests = describe('Layout AppBar component', () => {
+export const layoutAppBarTests: Mocha.Suite = describe('Layout AppBar component', () => {
 
     let c!: renderer.ReactTestRenderer
 
