@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { Provider } from 'react-redux'
 
+import { PictureInPicture } from '@material-ui/icons'
 import { expect } from 'chai'
-import { PictureInPicture } from 'material-ui-icons'
 import * as renderer from 'react-test-renderer'
 import { ToggleRedactionWidget } from '../../../src/components/document-widgets/ToggleRedaction'
 import { documentReceivedAction } from '../../../src/store/Document'
@@ -11,7 +11,7 @@ import { exampleDocumentData, useTestContext } from '../../viewercontext'
 /**
  * Toggle Redaction widget tests
  */
-export const toggleRedactionTests = describe('ToggleRedactionWidget component', () => {
+export const toggleRedactionTests: Mocha.Suite = describe('ToggleRedactionWidget component', () => {
 
     let c!: renderer.ReactTestRenderer
 
