@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { Provider } from 'react-redux'
 
+import { Dashboard } from '@material-ui/icons'
 import { expect } from 'chai'
-import { Dashboard } from 'material-ui-icons'
 import * as renderer from 'react-test-renderer'
 import { ToggleShapesWidget } from '../../../src/components/document-widgets/ToggleShapes'
 import { documentReceivedAction } from '../../../src/store/Document'
@@ -11,7 +11,7 @@ import { exampleDocumentData, useTestContext } from '../../viewercontext'
 /**
  * Toggle Shapes widget tests
  */
-export const toggleShapesWidgetTests = describe('ToggleShapesWidget component', () => {
+export const toggleShapesWidgetTests: Mocha.Suite = describe('ToggleShapesWidget component', () => {
 
     let c!: renderer.ReactTestRenderer
 
