@@ -1,16 +1,17 @@
 import * as React from 'react'
 import { Provider } from 'react-redux'
 
-import { IconButton, Menu } from 'material-ui'
+import { IconButton, Menu } from '@material-ui/core'
 import * as renderer from 'react-test-renderer'
-import { ZoomModeWidget } from '../../../src/components/document-widgets/ZoomMode'
+import { ZoomWidgetComponent } from '../../../src/components/document-widgets/ZoomMode'
 import { documentReceivedAction } from '../../../src/store/Document'
+import { defaultLocalization } from '../../../src/store/Localization'
 import { exampleDocumentData, useTestContext } from '../../viewercontext'
 
 /**
  * Zoom Mode widget tests
  */
-export const zoomModeWidgetTests = describe('ZoomModeWidget component', () => {
+export const zoomModeWidgetTests: Mocha.Suite = describe('ZoomModeWidget component', () => {
 
     let c!: renderer.ReactTestRenderer
 
@@ -24,43 +25,49 @@ export const zoomModeWidgetTests = describe('ZoomModeWidget component', () => {
 
             c = renderer.create(
                 <Provider store={ctx.store}>
-                    <ZoomModeWidget>
-                    </ZoomModeWidget>
+                    <ZoomWidgetComponent
+                        customZoomLevel={0}
+                        localization={defaultLocalization}
+                        setZoomLevel={() => ({} as any)}
+                        setZoomMode={() => ({} as any)}
+                        zoomMode="custom"
+                    >
+                    </ZoomWidgetComponent>
                 </Provider>)
         })
     })
 
     it('ZoomMenu open / close', () => {
-        useTestContext((ctx) => {
-            ctx.store.dispatch(documentReceivedAction(exampleDocumentData))
 
-            c = renderer.create(
-                <Provider store={ctx.store}>
-                    <ZoomModeWidget>
-                    </ZoomModeWidget>
-                </Provider>)
-            const zoomMenuButton = c.root.findByType(IconButton)
-            zoomMenuButton.props.onClick({currentTarget: undefined})
-            const zoomMenu = c.root.findByType(Menu)
-            zoomMenu.props.onClose()
-        })
+        c = renderer.create(
+            <ZoomWidgetComponent
+                customZoomLevel={0}
+                localization={defaultLocalization}
+                setZoomLevel={() => ({} as any)}
+                setZoomMode={() => ({} as any)}
+                zoomMode="custom">
+            </ZoomWidgetComponent>)
+        const zoomMenuButton = c.root.findByType(IconButton)
+        zoomMenuButton.props.onClick({ currentTarget: undefined })
+        const zoomMenu = c.root.findByType(Menu)
+        zoomMenu.props.onClose()
     })
 
-    //     ToDo: research on native ref testability
-    // it('set zoom mode to fit', async () => {
-    //     await useTestContextAsync(async (ctx) => {
-    //         ctx.store.dispatch(documentReceivedAction(exampleDocumentData))
-    //         c = renderer.create(
-    //             <Provider store={ctx.store}>
-    //                 <ZoomModeWidget>
-    //                 </ZoomModeWidget>
-    //             </Provider>)
-    //         const zoomMenuButton = c.root.findByType(IconButton)
-    //         zoomMenuButton.props.onClick({currentTarget: {...zoomMenuButton, children: []}})
-    //         await asyncDelay(1000)
-    //         const button = c.root.findAllByType(MenuItem)[0];
-    //         (button as any).props.onClick({})
-    //         expect(ctx.store.getState().sensenetDocumentViewer.viewer.zoomMode).to.be.eq('fit')
-    //     })
-    // })
+    it('set zoom mode to fit', async () => {
+        c = renderer.create(
+            <ZoomWidgetComponent
+                customZoomLevel={0}
+                localization={defaultLocalization}
+                setZoomLevel={() => ({} as any)}
+                setZoomMode={() => ({} as any)}
+                zoomMode="custom">
+            </ZoomWidgetComponent>)
+
+        /* ToDo: Fix me!!!
+        const zoomMenuButton = c.root.findByType(IconButton)
+        zoomMenuButton.props.onClick({ currentTarget: { ...zoomMenuButton, children: [] } })
+        await asyncDelay(1000)
+        const button = c.root.findAllByType(MenuItem)[0];
+        (button as any).props.onClick({})*/
+    })
 })
