@@ -1,8 +1,8 @@
 import * as React from 'react'
 import { Provider } from 'react-redux'
 
+import { RotateLeft, RotateRight } from '@material-ui/icons'
 import { expect } from 'chai'
-import { RotateLeft, RotateRight } from 'material-ui-icons'
 import * as renderer from 'react-test-renderer'
 import { RotatePageWidget } from '../../../src/components/page-widgets/RotatePage'
 import { documentReceivedAction } from '../../../src/store/Document'
@@ -12,7 +12,7 @@ import { exampleDocumentData, useTestContext } from '../../viewercontext'
 /**
  * RotatePage widget tests
  */
-export const rotatePageWidgetTests = describe('RotatePageWidget component', () => {
+export const rotatePageWidgetTests: Mocha.Suite = describe('RotatePageWidget component', () => {
 
     let c!: renderer.ReactTestRenderer
 
